requirements.txt
faker==25.8.0
psycopg2-binary==2.9.9
python-dotenv==0.17.1